wip: add hints to pixel jumble
Now only need an answering system
