+++ src/commands/pixel_jumble.ts
@@ -34,32 +34,38 @@ export default {
                 )
         ),
 
-    // for Rust:
-    // https://github.com/dylanleclair/pixels/blob/24688f58aae9b8b2b0fc3046687ebea0514c3061/src/main.rs#L48
     run: async ({ interaction, client }): Promise<void> => {
         const { type } = getOptions<{ type: MediaType }>(interaction.options, ["type"]);
 
         if (type != "ANIME" && type != "MANGA") throw new YuukoError(`Please specify either manga, or anime as your content type. (Yours was "${type}")`);
 
-        const vars = { type, userId: interaction.alID };
+        const msg = await interaction.deferReply({ withResponse: true });
+
+        const { data: viewer } = await graphQLRequest("Viewer", {}, interaction.ALtoken);
+        if (!viewer) throw new YuukoError("Couldn't fetch your profile using your token to get an entry");
+
+        const statistics = viewer.Viewer?.statistics;
+        const totalSize = type === "ANIME" ? statistics?.anime?.count : statistics?.manga?.count;;
+
+        if (!totalSize) throw new YuukoError(`We could not find enough entries in your list for the specified media type (got ${totalSize})`)
+
+        const vars = { type, userId: interaction.alID, chunk: Math.floor(Math.random() * totalSize) };
         const { data: { MediaListCollection: data } } = await graphQLRequest("PixelJumble", vars);
 
         if (!data || !data.lists || data.lists.length < 1) throw new YuukoError("Couldn't find any data from the user specified.", vars);
 
-        data.lists = data.lists.filter(list => list && list.name !== "Planning");
         const allMediaItems = data.lists
-            .filter((list): list is NonNullable<typeof list> => list != null && list.name !== "Planning")
-            .flatMap(list => list.entries);
+            .flatMap(list => list?.entries);
 
         const random = Math.floor(Math.random() * Math.floor(allMediaItems.length));
-        const entry = allMediaItems[random]!;
+        const mediaEntry = allMediaItems[random]!.media;
 
-        const title = SeriesTitle(entry.media?.title);
-        const coverImage = entry.media?.coverImage?.large || "https://i.imgur.com/Hx8474m.png";
+        if (!mediaEntry) throw new YuukoError("Error fetching media entry from list");
+
+        const title = SeriesTitle(mediaEntry.title);
+        const coverImage = mediaEntry.coverImage?.large || "https://i.imgur.com/Hx8474m.png";
 
         const lib = client.modules.getModule("modules");
-
-        const msg = await interaction.reply({ embeds: [{ description: "Preparing the game..." }], withResponse: true });
 
         const enc = new TextEncoder();
         const encodedImgUrl = enc.encode(coverImage);
@@ -85,8 +91,24 @@ export default {
         const customIDs = new Set(['guess', 'hint', 'forfeit']);
         const components = [new ActionRowBuilder<ButtonBuilder>().addComponents(buttonList)];
 
-        const collector = createButtonCollector(interaction, customIDs, msg);
-        let hints = "There are no hints.";
+        const collector = createButtonCollector(interaction, customIDs, msg, { time: 30_000 });
+
+        const possibleHints: string[] = [];
+
+        const { year: startYear, month: startMonth, day: startDay } = mediaEntry.startDate ?? {};
+        if (startYear !== null) possibleHints.push(`Has a start date of **${startYear}-${startMonth}-${startDay}**`);
+        if (mediaEntry.source) possibleHints.push(`Has a source of **${mediaEntry.source}**`);
+        if (mediaEntry.genres?.length) possibleHints.push(`Has genres **${mediaEntry.genres.join(", ")}**`);
+        if (mediaEntry.popularity) possibleHints.push(`Has a popularity of **${mediaEntry.popularity}**`);
+        if (mediaEntry.averageScore) possibleHints.push(`Has an average score of **${mediaEntry.averageScore}%**`);
+        if (mediaEntry.characters?.edges?.length) possibleHints.push(`Has a character named **${selectRandom(mediaEntry.characters.edges)!.name}**`);
+        if (mediaEntry.staff?.edges?.length) {
+            const randomStaff = selectRandom(mediaEntry.staff.edges)!;
+            possibleHints.push(`Has **${randomStaff.node?.name?.full}** who worked as **${randomStaff.role}**`);
+        }
+        if (mediaEntry.countryOfOrigin) possibleHints.push(`Has an origin of ${mediaEntry.countryOfOrigin}`);
+
+        let hints = `- ${selectRandom(possibleHints)}\n- ${selectRandom(possibleHints)}\n- ${selectRandom(possibleHints)}\n`;
 
         const embed = new EmbedBuilder()
             .setTitle(`Pixel Jumble - Guess the ${type.toLowerCase()}`)
@@ -94,23 +116,43 @@ export default {
 
         await interaction.editReply({ files: [attachment], components, embeds: [embed] })
 
+        let forfeit = false;
         collector?.on("collect", async (i) => {
             if (!i.isButton()) return;
 
             switch (i.customId) {
                 case "hint":
-                    pixelationLevel = Math.max(1, pixelationLevel - 1.5);
+                    pixelationLevel = Math.max(0, pixelationLevel - 1.5);
 
-                    if (pixelationLevel <= 1 || originalImgPtr === null) break;
+                    if (pixelationLevel < 1 || originalImgPtr === null) {
+                        i.deferUpdate();
+                        break;
+                    };
+
                     if (!i.deferred) await i.deferUpdate();
 
                     [pixelatedImgPtr, pixelatedImgBufferSize] = pixelateImage(lib, originalImgPtr, pixelationLevel);
                     attachment = getAttachment(pixelatedImgPtr, pixelatedImgBufferSize);
 
-                    await interaction.editReply({ files: [attachment] })
+                    if (pixelationLevel === 5.5) hints += `- ${selectRandom(possibleHints)}\n`; // First hint
+                    else if (pixelationLevel === 4) hints += `- ${selectRandom(possibleHints)}\n`; // Second hint
+                    else if (pixelationLevel === 2.5) hints += `- ${selectRandom(possibleHints)}\n`; // Third hint
+                    else if (pixelationLevel === 1) hints += `- **${scrambleSentence(title)}**`; // Fourth hint
+
+                    embed.setDescription(hints);
+
+                    if (pixelationLevel === 1) {
+                        buttonList[1]?.setDisabled(true)
+                        components[0]!.setComponents(buttonList);
+                        await interaction.editReply({ files: [attachment], components, embeds: [embed] })
+                        break;
+                    }
+
+                    await interaction.editReply({ files: [attachment], embeds: [embed] })
 
                     break;
                 case "forfeit":
+                    forfeit = true;
                     collector.stop();
                     break;
             }
@@ -120,7 +162,7 @@ export default {
             [pixelatedImgPtr, pixelatedImgBufferSize] = pixelateImage(lib, originalImgPtr, 1);
             attachment = getAttachment(pixelatedImgPtr, pixelatedImgBufferSize);
 
-            embed.setDescription(`${hints}\n\n**${interaction.user.displayName}** gave up!\nAnswer was ${title}`)
+            embed.setDescription(`${hints}\n\n**${interaction.user.displayName}** ${forfeit ? "gave up" : "failed to guess in time"}!\nAnswer was **${title}**`)
             await interaction.editReply({ files: [attachment], components: [], embeds: [embed] })
 
             if (originalImgPtr) {
@@ -162,3 +204,18 @@ function getAttachment(imgPtr: Pointer, bufferSize: number) {
 
     return attachment;
 };
+
+function scrambleSentence(input: string) {
+    const scrambleWord = (word: string) => word
+        .split("")
+        .sort(() => Math.random() - 0.5)
+        .join("");
+
+    return input.split(" ").map(scrambleWord).join(" ").toLocaleUpperCase();
+}
+
+function selectRandom<T>(source: T[]) {
+    if (source.length === 0) return undefined;
+    const idx = Math.floor(Math.random() * source.length);
+    return source.splice(idx, 1)[0];
+}
