+++ src/graphQL/types/index.ts
@@ -4736,10 +4736,11 @@ export type MangaQuery = { __typename?: 'Query', Media?: { __typename?: 'Media',
 export type PixelJumbleQueryVariables = Exact<{
   userId?: InputMaybe<Scalars['Int']['input']>;
   type?: InputMaybe<MediaType>;
+  chunk?: InputMaybe<Scalars['Int']['input']>;
 }>;
 
 
-export type PixelJumbleQuery = { __typename?: 'Query', MediaListCollection?: { __typename?: 'MediaListCollection', lists?: Array<{ __typename?: 'MediaListGroup', name?: string | null, entries?: Array<{ __typename?: 'MediaList', media?: { __typename?: 'Media', title?: { __typename?: 'MediaTitle', romaji?: string | null, english?: string | null, native?: string | null } | null, coverImage?: { __typename?: 'MediaCoverImage', large?: string | null, medium?: string | null } | null } | null } | null> | null } | null> | null } | null };
+export type PixelJumbleQuery = { __typename?: 'Query', MediaListCollection?: { __typename?: 'MediaListCollection', lists?: Array<{ __typename?: 'MediaListGroup', name?: string | null, entries?: Array<{ __typename?: 'MediaList', media?: { __typename?: 'Media', averageScore?: number | null, genres?: Array<string | null> | null, popularity?: number | null, countryOfOrigin?: any | null, source?: MediaSource | null, title?: { __typename?: 'MediaTitle', romaji?: string | null, english?: string | null, native?: string | null } | null, coverImage?: { __typename?: 'MediaCoverImage', large?: string | null } | null, startDate?: { __typename?: 'FuzzyDate', day?: number | null, month?: number | null, year?: number | null } | null, characters?: { __typename?: 'CharacterConnection', edges?: Array<{ __typename?: 'CharacterEdge', name?: string | null } | null> | null } | null, staff?: { __typename?: 'StaffConnection', edges?: Array<{ __typename?: 'StaffEdge', role?: string | null, node?: { __typename?: 'Staff', name?: { __typename?: 'StaffName', full?: string | null } | null } | null } | null> | null } | null } | null } | null> | null } | null> | null } | null };
 
 export type RecentChartQueryVariables = Exact<{
   user?: InputMaybe<Scalars['String']['input']>;
@@ -4812,4 +4813,4 @@ export type UserCardQuery = { __typename?: 'Query', User?: { __typename?: 'User'
 export type ViewerQueryVariables = Exact<{ [key: string]: never; }>;
 
 
-export type ViewerQuery = { __typename?: 'Query', Viewer?: { __typename?: 'User', name: string, id: number } | null };
+export type ViewerQuery = { __typename?: 'Query', Viewer?: { __typename?: 'User', id: number, name: string, statistics?: { __typename?: 'UserStatisticTypes', anime?: { __typename?: 'UserStatistics', count: number } | null, manga?: { __typename?: 'UserStatistics', count: number } | null } | null } | null };
