+++ src/graphQL/types/queries.ts
@@ -13,8 +13,8 @@ export default {
   "Character": "query Character($charName: String) {\n  Character(search: $charName) {\n    name {\n      full\n    }\n    age\n    description\n    siteUrl\n    image {\n      large\n    }\n    bloodType\n    dateOfBirth {\n      year\n      month\n      day\n    }\n    gender\n    favourites\n    media {\n      nodes {\n        siteUrl\n        title {\n          romaji\n          native\n          english\n        }\n      }\n    }\n  }\n}\n",
   "Manga": "query Manga($query: String) {\n  Media(search: $query, type: MANGA, sort: SEARCH_MATCH) {\n    id\n    description\n    coverImage {\n      large\n      medium\n    }\n    title {\n      romaji\n      english\n      native\n    }\n    format\n    chapters\n    source\n    synonyms\n    volumes\n    genres\n    id\n    siteUrl\n    meanScore\n    startDate {\n      year\n      month\n      day\n    }\n    endDate {\n      year\n      month\n      day\n    }\n    bannerImage\n    mediaListEntry {\n      status\n      progress\n      score(format: POINT_10_DECIMAL)\n      notes\n      user {\n        name\n        mediaListOptions {\n          scoreFormat\n        }\n      }\n    }\n  }\n}\n",
   "UserCard": "query UserCard($username: String) {\n  User(name: $username) {\n    id\n    name\n    avatar {\n      large\n      medium\n    }\n    bannerImage\n    siteUrl\n    createdAt\n    statistics {\n      anime {\n        count\n        meanScore\n      }\n      manga {\n        count\n        meanScore\n      }\n    }\n  }\n}\n",
-  "PixelJumble": "query PixelJumble($userId: Int, $type: MediaType) {\n    MediaListCollection(userId: $userId, type: $type) {\n        lists {\n            entries {\n                media {\n                    title {\n                        romaji\n                        english\n                        native\n                    }\n                    coverImage {\n                        large\n                        medium\n                    }\n                }\n            }\n            name\n        }\n    }\n}\n",
+  "PixelJumble": "query PixelJumble($userId: Int, $type: MediaType, $chunk: Int) {\n    MediaListCollection(userId: $userId, type: $type, chunk: $chunk, perChunk: 1, status_not: PLANNING) {\n        lists {\n            entries {\n                media {\n                    title {\n                        romaji\n                        english\n                        native\n                    }\n                    coverImage {\n                        large\n                    }\n                    averageScore\n                    genres\n                    popularity\n                    countryOfOrigin\n                    source\n                    startDate {\n                        day\n                        month\n                        year\n                    }\n                    characters(perPage: 1) {\n                        edges {\n                            name\n                        }\n                    }\n                    staff(perPage: 1) {\n                        edges {\n                            node {\n                                name {\n                                    full\n                                }\n                            }\n                            role\n                        }\n                    }\n                }\n            }\n            name\n        }\n    }\n}\n",
   "SaveMediaList": "mutation SaveMediaListEntry($mediaid: Int, $status: MediaListStatus, $score: Float, $progress: Int, $hide: Boolean, $lists: [String], $private: Boolean) {\n  SaveMediaListEntry(mediaId: $mediaid, status: $status, score: $score, progress: $progress, hiddenFromStatusLists: $hide, customLists: $lists, private: $private) {\n    mediaId\n    media {\n      bannerImage\n      type\n      title {\n        userPreferred\n      }\n    }\n    user {\n      name\n    }\n    status\n  }\n}\n",
   "Activity": "query Activity($userid: Int) {\n  Activity(userId: $userid, sort: ID_DESC, type_not: MESSAGE) {\n    __typename\n    ... on TextActivity {\n      text\n      siteUrl\n      createdAt\n      user {\n        name\n        avatar {\n          large\n        }\n      }\n      likeCount\n      replyCount\n      replies {\n        text\n        user {\n          name\n        }\n      }\n    }\n    ... on ListActivity {\n      user {\n        name\n        avatar {\n          large\n        }\n      }\n      likeCount\n      replies {\n        text\n        user {\n          name\n        }\n      }\n      replyCount\n      status\n      progress\n      media {\n        siteUrl\n        bannerImage\n        title {\n          romaji\n          native\n          english\n        }\n        coverImage {\n          extraLarge\n          large\n          medium\n        }\n      }\n      createdAt\n      siteUrl\n    }\n    ... on MessageActivity {\n      createdAt\n    }\n  }\n}",
-  "Viewer": "query Viewer {\n  Viewer {\n    name\n    id\n  }\n}\n"
+  "Viewer": "query Viewer {\n    Viewer {\n        statistics {\n            anime {\n                count\n            }\n            manga {\n                count\n            }\n        }\n        id\n        name\n    }\n}\n"
 } as const
