+++ src/modules/Cargo.toml
@@ -5,13 +5,17 @@ edition = "2024"
 
 [dependencies]
 serde_json = "1.0.140"
-serde = { version = "1.0.219", features = ["derive"]}
+serde = { version = "1.0.219", features = ["derive"] }
 imageproc = "0.25.0"
-image = { version = "0.25.6", default-features = false, features = ["png", "jpeg"] }
+image = { version = "0.25.6", default-features = false, features = [
+    "png",
+    "jpeg",
+] }
 reqwest = { version = "0.12", features = ["rustls-tls"] }
 ab_glyph = "0.2.29"
 anyhow = "1.0.98"
 tokio = { version = "1", features = ["full"] }
+once_cell = "1.21.3"
 
 [profile.release]
 strip = true
